requirements.txt
tensorflow
numpy
pandas
opencv-python
matplotlib
scikit-learn
flask
pillow
gunicorn
